+++ tencentcloud/ckafka/v20190819/models.go
@@ -1532,7 +1532,7 @@ type CreateInstancePreRequestParams struct {
 	// 实例名称，是一个不超过 64 个字符的字符串，必须以字母为首字符，剩余部分可以包含字母、数字和横划线(-)
 	InstanceName *string `json:"InstanceName,omitempty" name:"InstanceName"`
 
-	// 可用区
+	// 可用区，购买多可用区实例时，填写ZoneIds.N字段中的任意一个值
 	ZoneId *int64 `json:"ZoneId,omitempty" name:"ZoneId"`
 
 	// 预付费购买时长，例如 "1m",就是一个月
@@ -1581,7 +1581,7 @@ type CreateInstancePreRequestParams struct {
 	// 跨可用区，zoneIds必填
 	MultiZoneFlag *bool `json:"MultiZoneFlag,omitempty" name:"MultiZoneFlag"`
 
-	// 可用区列表
+	// 可用区列表，购买多可用区实例时为必填项
 	ZoneIds []*int64 `json:"ZoneIds,omitempty" name:"ZoneIds"`
 }
 
@@ -1591,7 +1591,7 @@ type CreateInstancePreRequest struct {
 	// 实例名称，是一个不超过 64 个字符的字符串，必须以字母为首字符，剩余部分可以包含字母、数字和横划线(-)
 	InstanceName *string `json:"InstanceName,omitempty" name:"InstanceName"`
 
-	// 可用区
+	// 可用区，购买多可用区实例时，填写ZoneIds.N字段中的任意一个值
 	ZoneId *int64 `json:"ZoneId,omitempty" name:"ZoneId"`
 
 	// 预付费购买时长，例如 "1m",就是一个月
@@ -1640,7 +1640,7 @@ type CreateInstancePreRequest struct {
 	// 跨可用区，zoneIds必填
 	MultiZoneFlag *bool `json:"MultiZoneFlag,omitempty" name:"MultiZoneFlag"`
 
-	// 可用区列表
+	// 可用区列表，购买多可用区实例时为必填项
 	ZoneIds []*int64 `json:"ZoneIds,omitempty" name:"ZoneIds"`
 }