+++ tencentcloud/mps/v20190612/models.go
@@ -5432,6 +5432,9 @@ type EditMediaFileInfo struct {
 type EditMediaOutputConfig struct {
 	// 封装格式，可选值：mp4、hls、mov、flv、avi。默认是 mp4。
 	Container *string `json:"Container,omitempty" name:"Container"`
+
+	// 剪辑模式，可选值 normal、fast。默认是精确剪辑 normal
+	Type *string `json:"Type,omitempty" name:"Type"`
 }
 
 // Predefined struct for user